+++ js/src/main/scala/com/raquo/domtypes/jsdom/defs/events/TypedTargetEvent.scala
@@ -6,6 +6,6 @@ import scala.scalajs.js
 
 /** Represents an event for which we know what exact type the `target` is. */
 @js.native
-trait ElementTargetEvent[+E <: dom.html.Element] extends dom.Event {
+trait TypedTargetEvent[+E <: dom.EventTarget] extends dom.Event {
   override def target: E = js.native
 }