+++ js/src/test/scala/com/raquo/domtypes/CompileTest.scala
@@ -8,7 +8,7 @@ import com.raquo.domtypes.generic.defs.props.Props
 import com.raquo.domtypes.generic.defs.reflectedAttrs.ReflectedAttrs
 import com.raquo.domtypes.generic.defs.styles.{Styles, Styles2}
 import com.raquo.domtypes.generic.keys.{Attr, EventProp, Prop, Style}
-import com.raquo.domtypes.jsdom.defs.eventProps.{ClipboardEventProps, ErrorEventProps, FormEventProps, KeyboardEventProps, MediaEventProps, MiscellaneousEventProps, MouseEventProps, WindowEventProps}
+import com.raquo.domtypes.jsdom.defs.eventProps.{ClipboardEventProps, ErrorEventProps, HTMLElementEventProps, FormEventProps, KeyboardEventProps, MediaEventProps, MiscellaneousEventProps, MouseEventProps, WindowOnlyEventProps}
 import com.raquo.domtypes.jsdom.defs.tags.{DocumentTags, EmbedTags, FormTags, GroupingTags, MiscTags, SectionTags, TableTags, TextTags}
 import org.scalajs.dom
 
@@ -43,7 +43,7 @@ class CompileTest {
     with MediaEventProps[EventProp]
     with MiscellaneousEventProps[EventProp]
     with MouseEventProps[EventProp]
-    with WindowEventProps[EventProp]
+    with WindowOnlyEventProps[EventProp]
     // Props
     with Props[Prop]
     // Reflected Attrs
@@ -67,4 +67,8 @@ class CompileTest {
     with CanonicalPropBuilder
     with SomeStyleBuilders
     with SomeTagBuilder
+
+  object SpecificEventProps
+    extends HTMLElementEventProps[EventProp]
+    with CanonicalEventPropBuilder[dom.Event]
 }