+++ shared/src/main/scala/com/raquo/domtypes/generic/defs/eventProps/FormEventProps.scala
@@ -3,15 +3,18 @@ package com.raquo.domtypes.generic.defs.eventProps
 import com.raquo.domtypes.generic.builders.EventPropBuilder
 
 /**
+  * @tparam DomElementTargetEvent
+  *            An event that has an Element as `target`.
+  *            This event type has no corresponding type in JS DOM. See our own `TypedTargetEvent` trait.
   * @tparam DomHtmlElementTargetEvent
   *            An event that has an HTMLElement as `target`.
-  *            This event type has no corresponding type in JS DOM. See our own `ElementTargetEvent` trait.
+  *            This event type has no corresponding type in JS DOM. See our own `TypedTargetEvent` trait.
   * @tparam DomFormElementTargetEvent
   *            An event that has an HTMLFormElement as `target`.
-  *            This event type has no corresponding type in JS DOM. See our own `ElementTargetEvent` trait.
+  *            This event type has no corresponding type in JS DOM. See our own `TypedTargetEvent` trait.
   * @tparam DomInputElementTargetEvent
   *            An event that has an HTMLInputElement as `target`.
-  *            This event type has no corresponding type in JS DOM. See our own `ElementTargetEvent` trait.
+  *            This event type has no corresponding type in JS DOM. See our own `TypedTargetEvent` trait.
   *            - Strictly speaking, this type is not applicable to `onInput` because the latter
   *              can be fired on any HTMLElement that has `contentEditable` mode enabled.
   *            - Similarly for `onChange` and `onSelect` – these could also fire on an `HTMLTextAreaElement`.
@@ -19,7 +22,7 @@ import com.raquo.domtypes.generic.builders.EventPropBuilder
   *            DOM InputEvent https://developer.mozilla.org/en-US/docs/Web/API/InputEvent
   *            Note: This type is not implemented in scala-js-dom
   */
-trait FormEventProps[EP[_ <: DomEvent], DomEvent, DomFocusEvent <: DomEvent, DomInputEvent <: DomEvent, DomHtmlElementTargetEvent <: DomEvent, DomFormElementTargetEvent <: DomEvent, DomInputElementTargetEvent <: DomEvent] { this: EventPropBuilder[EP, DomEvent] =>
+trait FormEventProps[EP[_ <: DomEvent], DomEvent, DomFocusEvent <: DomEvent, DomInputEvent <: DomEvent, DomElementTargetEvent <: DomEvent, DomHtmlElementTargetEvent <: DomEvent, DomFormElementTargetEvent <: DomEvent, DomInputElementTargetEvent <: DomEvent] { this: EventPropBuilder[EP, DomEvent] =>
 
   /**
     * The change event is fired for input, select, and textarea elements
@@ -48,7 +51,7 @@ trait FormEventProps[EP[_ <: DomEvent], DomEvent, DomFocusEvent <: DomEvent, Dom
     *
     * MDN
     */
-  lazy val onBlur: EP[DomFocusEvent with DomHtmlElementTargetEvent] = eventProp("blur")
+  lazy val onBlur: EP[DomFocusEvent with DomElementTargetEvent] = eventProp("blur")
 
 
   /**
@@ -56,7 +59,7 @@ trait FormEventProps[EP[_ <: DomEvent], DomEvent, DomFocusEvent <: DomEvent, Dom
     *
     * MDN
     */
-  lazy val onFocus: EP[DomFocusEvent with DomHtmlElementTargetEvent] = eventProp("focus")
+  lazy val onFocus: EP[DomFocusEvent with DomElementTargetEvent] = eventProp("focus")
 
   /**
     * The submit event is fired when the user clicks a submit button in a form
@@ -76,7 +79,7 @@ trait FormEventProps[EP[_ <: DomEvent], DomEvent, DomFocusEvent <: DomEvent, Dom
   /**
     * Script to be run when an element is invalid
     */
-  lazy val onInvalid: EP[DomHtmlElementTargetEvent] = eventProp("invalid")
+  lazy val onInvalid: EP[DomElementTargetEvent] = eventProp("invalid")
 
   /**
     * Fires when the user writes something in a search field (for <input="search">)